Move import_ds() to its own module.
